tutorial 6: user input
This tutorial had to take a pretty big departure from the guided path
because the golang.org/x/mobile/app framework is primarily focused on
touch interfaces.

None the less, in the end you end up with a similar result from
completing the exercises, in which the camera translates along a
circular track instead of freely in space.  The is no direct control
over the camera angle.

It could be another fun exercise to use sensor input to get the camera
angle relative to center.  There are a number of design problems in
doing such a thing, I think.