tutorial 9: VBO Indexing
Something about the ELEMENT_ARRAY_BUFFER is breaking the debug.FPS
gauge.  It happens that this tutorial suggests you log frame latency to
stderr, which works fine as a workaround to a broken graphical meter.

The specific effect of attempting to combine the two things is that the
cube gets textured from the FPS meter somehow.  This may be possible to
work around by looking into the implementation of the debug package and
the glutil package.